requirements.txt
pyserial==3.3
python-telegram-bot
